requirements.txt
onvif-zeep==0.2.12
pillow==9.5.0
pynput==1.7.6
pystray==0.19.4